requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.3
django-environ==0.11.2
djangorestframework==3.15.1
gunicorn==21.2.0
packaging==24.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
